requirements.txt
Pillow
torch
torchvision 
tensorboard